+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Add `Value` interface to represent a value in wrappers
 - Add `Entity` interface to represent an entity
 - Add `Validatable` interface to represent a validatable object
+`reflection` module:
+- `decodeObject()` now auto-validates the object if it implements `Validatable`
 ### Fixed Bugs
 `core-mit` module:
 - Wrapping of `WrappedException` now returns the original `WrappedException`